:truck: pricing 페이지 fsd 적용 #15
